refactor(api): host the MCP handler at app/mcp; versioned path becomes the alias
The implementation now lives at the canonical /mcp route;
/api/v2/agent/[transport] is a thin re-export kept for
already-configured clients.
